Further Implementation
I had a few problems but ended up fixing them with the help of forums. One problem was within the controller. A specific type of code no longer works, within the ruby version i am using. To resolve this it was a simple changing .for to .permit for the sign in feature.